# swagger.yaml
openapi: 3.0.4
info:
  title: Киносервис api
  description: api для управления пользователями, подписками и контентом киносервиса.
  version: 1.0.0
servers:
  - url: http://localhost:3005/api
...
paths:
...
  /admin/users:
    get:
      summary: Получить список пользователей
      description: >
        **Метод для получения списка пользователей.** Возвращает список пользователей с возможностью фильтрации по email или username.
        **Параметры:**
        - `page`: Номер страницы (по умолчанию 1).
        - `limit`: Количество пользователей на странице (по умолчанию 25, 0 - все).
        - `search`: Строка для поиска по email или username (по умолчанию пустая строка).
      operationId: getAdminUsers
      tags:
        - admin
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Номер страницы
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Количество пользователей на странице
          schema:
            type: integer
            default: 25
            minimum: 0
        - name: search
          in: query
          description: Строка для поиска по email или username
          schema:
            type: string
            default: ""
      responses:
        "200":
          description: Успешно получен список пользователей
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAdminUsersResponse"
        "401":
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: "unauthorized"
                message: "Not authenticated"
        "403":
          description: Отсутствуют права доступа
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: "forbidden"
                message: "Доступ запрещен"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: "internal_server_error"
                message: "Ошибка сервера при получении списка пользователей"
  /admin/user/change/{userId}:
    put:
      summary: Изменить данные пользователя
      description: >
        **Метод для изменения данных пользователя администратором.** Принимает ID пользователя и данные для обновления.
        **Параметры:**
        - `userId`: ID пользователя, данные которого нужно изменить.
        - `request_data`: Данные для обновления (username, email, wallet, currentSubscription).
      operationId: adminChangeUser
      tags:
        - admin
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя
          schema:
            type: string
            format: ObjectId
      requestBody:
        description: Данные для обновления пользователя
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminChangeUserRequest"
      responses:
        "200":
          description: Данные пользователя успешно обновлены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminChangeUserResponse"
        "400":
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: "bad_request"
                message: "Неверный формат email или имя пользователя слишком короткое"
        "401":
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: "unauthorized"
                message: "Not authenticated"
        "403":
          description: Отсутствуют права доступа
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: "forbidden"
                message: "Доступ запрещен"
        "404":
          description: Пользователь или план подписки не найдены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: "not_found"
                message: "Пользователь не найден"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: "internal_server_error"
                message: "Ошибка сервера при обновлении пользователя"
  /admin/user/delete/{userId}:
    delete:
      summary: Удалить пользователя
      description: >
        **Метод для удаления пользователя администратором.** Принимает ID пользователя и удаляет его из базы данных.
        **Параметры:**
        - `userId`: ID пользователя, которого нужно удалить.
      operationId: adminDeleteUser
      tags:
        - admin
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя
          schema:
            type: string
            format: ObjectId
      responses:
        "200":
          description: Пользователь успешно удален
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDeleteUserResponse"
        "401":
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: "unauthorized"
                message: "Not authenticated"
        "403":
          description: Отсутствуют права доступа
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: "forbidden"
                message: "Доступ запрещен"
        "404":
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: "not_found"
                message: "Пользователь не найден"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: "internal_server_error"
                message: "Ошибка сервера при удалении пользователя"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          description: Код ошибки
        message:
          type: string
          description: Сообщение об ошибке
...
    SubscriptionPlanResponse:
      type: array
      items:
        type: object
        properties:
          _id:
            type: string
            description: Уникальный идентификатор тарифного плана.
          name:
            type: string
            description: Название тарифного плана.
          price:
            type: number
            description: Стоимость подписки.
          renewalPeriod:
            type: number
            description: Период действия подписки в днях.
          features:
            type: array
            description: Список преимуществ и функций, включенных в план.
            items:
              type: string
          createdAt:
            type: string
            format: date-time
            description: Дата и время создания тарифного плана.
          updatedAt:
            type: string
            format: date-time
            description: Дата и время последнего обновления тарифного плана.
      example:
        - _id: 67f437687b06d9a11720a6ce
          name: Базовый
          price: 0
          renewalPeriod: 30
          features:
            - Full HD качество
            - 1 устройство
            - С рекламой
          createdAt: "2025-04-07T20:36:56.674000"
          updatedAt: "2025-04-18T20:41:39.084000"
        - _id: 67f437687b06d9a11720a6d1
          name: Популярный
          price: 899
          renewalPeriod: 31
          features:
            - 4K Ultra HD + HDR
            - До 5 устройств
            - Без рекламы
            - Оффлайн просмотр
          createdAt: "2025-04-07T20:36:56.727000"
          updatedAt: "2025-05-23T22:22:21.536000"
        - _id: 67f437687b06d9a11720a6d4
          name: Премиум+
          price: 1199
          renewalPeriod: 31
          features:
            - 4K Ultra HD + HDR + Dolby Vision
            - До 7 устройств
            - Без рекламы + ранний доступ
            - Оффлайн-просмотр + эксклюзивы
          createdAt: "2025-04-07T20:36:56.740000"
          updatedAt: "2025-04-07T20:36:56.740000"
...
    UserCore:
      type: object
      properties:
        _id:
          type: string
          example: 688be24eb1ab066a2279502b
        username:
          type: string
          example: new_user
        email:
          type: string
          example: user@example.com
        avatar:
          type: string
          example: /defaults/default-avatar.png
        notifications:
          type: object
          properties:
            email:
              type: boolean
              example: true
            push:
              type: boolean
              example: false
            newsletter:
              type: boolean
              example: false
        createdAt:
          type: string
          format: date-time
          example: 2025-07-31T21:38:22.163000
        updatedAt:
          type: string
          format: date-time
          example: 2025-07-31T21:38:22.163000
        role:
          type: string
          example: user
        currentSubscription:
          type: object
          properties:
            planId:
              type: object
              description: >
                Обратите внимание: в примере `planId` - пустой объект,
                но обычно он должен содержать строку с ID плана.
            startDate:
              type: string
              format: date-time
              example: 2025-07-31T21:38:22.163000
            endDate:
              type: string
              nullable: true
              example: null
            isActive:
              type: boolean
              example: true
            autoRenew:
              type: boolean
              example: false
            adminNote:
              type: string
              nullable: true
              example: null
            plan:
              type: object
              properties:
                id:
                  type: string
                  example: 67f437687b06d9a11720a6ce
                name:
                  type: string
                  example: Базовый
                price:
                  type: number
                  example: 0
                features:
                  type: array
                  items:
                    type: string
                  example:
                    - Full HD качество
                    - 1 устройство
                    - С рекламой
                renewalPeriod:
                  type: number
                  example: 30
                createdAt:
                  type: string
                  format: date-time
                  example: 2025-04-07T20:36:56.674000
                updatedAt:
                  type: string
                  format: date-time
                  example: 2025-04-18T20:41:39.084000
        wallet:
          type: object
          properties:
            balance:
              type: number
              example: 0
            transactionIds:
              type: array
              items:
                type: string
              example: []
        history:
          type: array
          items:
            type: object
          example: []
...
    GetAdminUsersResponse:
      type: object
      properties:
        success:
          type: boolean
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserListItem"
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer
    AdminUserListItem:
      type: object
      properties:
        _id:
          type: string
          format: ObjectId
        username:
          type: string
        email:
          type: string
          format: email
        currentSubscription:
          $ref: "#/components/schemas/AdminUserSubscription"
        wallet:
          $ref: "#/components/schemas/AdminUserWallet"
        createdAt:
          type: string
          format: date-time
    AdminUserSubscription:
      type: object
      nullable: true
      properties:
        planId:
          type: string
          format: ObjectId
          nullable: true
        startDate:
          type: string
          format: date-time
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
        autoRenew:
          type: boolean
        adminNote:
          type: string
          nullable: true
        plan:
          $ref: "#/components/schemas/SubscriptionPlanResponse"
          nullable: true
    AdminUserWallet:
      type: object
      properties:
        balance:
          type: number
    AdminChangeUserRequest:
      type: object
      properties:
        username:
          type: string
          description: Имя пользователя.
          example: new_user
        email:
          type: string
          format: email
          description: Email пользователя.
          example: user@example.com
        wallet:
          type: object
          description: Обновление данных кошелька.
          properties:
            balance:
              type: number
              description: Новый баланс кошелька.
              example: 11111
        currentSubscription:
          type: object
          description: Обновление данных подписки.
          properties:
            planId:
              type: string
              description: ID нового плана подписки.
              example: 67f437687b06d9a11720a6ce
            isActive:
              type: boolean
              description: Статус активности подписки.
              example: true
            autoRenew:
              type: boolean
              description: Статус автоматического продления.
              example: false
            adminNote:
              type: string
              description: Примечание администратора.
              example: Comment
            endDate:
              type: string
              format: date
              description: Дата окончания подписки.
              example: 2222-02-12
    AdminChangeUserResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/UserCore"
    AdminDeleteUserResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
...